requirements.txt
altgraph==0.17.2
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
beautifulsoup4==4.10.0
black==22.1.0
blis==0.7.8
bs4==0.0.1
catalogue==2.0.7
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cv==1.0.0
cycler==0.11.0
cymem==2.0.6
debugpy==1.5.1
decorator==5.1.1
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.1
django-cors-headers==3.11.0
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
djongo==1.3.6
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.4.0/en_core_web_lg-3.4.0-py3-none-any.whl
entrypoints==0.4
executing==0.8.2
filelock==3.4.2
Flask==2.0.3
fonttools==4.29.1
future==0.18.2
idna==3.3
install==1.3.5
ipykernel==6.9.1
ipython==8.0.1
itsdangerous==2.1.0
jedi==0.18.1
Jinja2==3.0.3
jupyter-client==7.1.2
jupyter-core==4.9.2
kiwisolver==1.3.2
langcodes==3.3.0
MarkupSafe==2.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
mongoengine==0.24.1
MouseInfo==0.1.3
murmurhash==1.0.7
mypy-extensions==0.4.3
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.1.0
nest-asyncio==1.5.4
numpy==1.22.2
opencv-python==4.5.5.62
packaging==21.3
pandas==1.4.1
parso==0.8.3
pathspec==0.9.0
pathy==0.6.2
pefile==2022.5.30
pickleshare==0.7.5
Pillow==9.0.1
pipenv==2022.1.8
platformdirs==2.4.1
preshed==3.0.6
prompt-toolkit==3.0.28
psycopg2==2.9.3
pure-eval==0.2.2
PyAudio @ file:///C:/Users/Gaurav/pipwin/PyAudio-0.2.11-cp39-cp39-win_amd64.whl
PyAutoGUI==0.9.53
pydantic==1.9.1
PyGetWindow==0.0.9
Pygments==2.11.2
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.7
PyJWT==2.3.0
pymongo==3.12.3
PyMsgBox==1.0.9
pyparsing==3.0.7
pyperclip==1.8.2
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.28
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2021.3
pywhatkit==5.3
pywin32==303
pywin32-ctypes==0.2.0
pyzmq==22.3.0
requests==2.27.1
screeninfo==0.8
six==1.16.0
smart-open==5.2.1
soupsieve==2.3.1
spacy==3.4.0
spacy-legacy==3.0.9
spacy-loggers==1.0.3
sqlparse==0.2.4
srsly==2.4.4
stack-data==0.2.0
thinc==8.1.0
tomli==2.0.1
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typer==0.4.2
typing_extensions==4.1.1
tzdata==2021.5
urllib3==1.26.8
virtualenv==20.13.0
virtualenv-clone==0.5.7
wasabi==0.9.1
wcwidth==0.2.5
Werkzeug==2.0.3
whitenoise==5.3.0
wikipedia==1.4.0
xlwt==1.3.0
